kernel-hal: merge commom serial code
