+++ Cargo.toml
@@ -1,9 +1,10 @@
 [workspace]
 members = [
     "zircon-object",
+    "zircon-syscall",
     "zircon-loader",
     "linux-object",
-    "zircon-syscall",
+    "linux-syscall",
     "linux-loader",
     "kernel-hal",
 ]
@@ -12,5 +13,4 @@ exclude = [
     "zircon-user",
     "zCore",
     "rboot",
-    "linux-syscall",
 ]
