+++ kernel-hal/src/bare/arch/riscv/interrupt.rs
@@ -2,7 +2,7 @@ use alloc::boxed::Box;
 use riscv::register::{sie, sstatus};
 use spin::Mutex;
 
-use super::{consts, plic, serial, trap, uart};
+use super::{plic, trap};
 use crate::utils::irq_manager::IrqManager;
 
 // IRQ
@@ -18,22 +18,13 @@ lazy_static::lazy_static! {
 #[allow(dead_code)]
 fn init_soft() {
     unsafe { sie::set_ssoft() };
-    sbi_println!("+++ setup soft int! +++");
+    info!("+++ setup soft int! +++");
 }
 
 fn init_ext() {
     unsafe { sie::set_sext() };
     plic::init();
-    sbi_println!("+++ Setting up PLIC +++");
-}
-
-fn init_uart() {
-    uart::init(consts::UART_BASE);
-
-    //但当没有SBI_CONSOLE_PUTCHAR时，却为什么不行？
-    serial::uart_print_fmt(format_args!("UART output testing\n\r"));
-
-    sbi_println!("+++ Setting up UART interrupts +++");
+    info!("+++ Setting up PLIC +++");
 }
 
 fn init_irq() {
@@ -47,10 +38,9 @@ fn init_irq() {
 
 pub(super) fn init() {
     unsafe { sstatus::set_sie() };
-    init_uart();
     init_ext();
     init_irq();
-    sbi_println!("+++ setup interrupt +++");
+    info!("+++ setup interrupt OK +++");
 }
 
 hal_fn_impl! {
